+++ libdmabufinfo/include/dmabufinfo/dmabuf_sysfs_stats.h
@@ -32,7 +32,7 @@ namespace dmabufinfo {
 struct DmabufInfo {
     unsigned long inode;
     std::string exp_name;
-    unsigned int size;
+    uint64_t size;
 };
 
 struct DmabufTotal {
@@ -85,7 +85,7 @@ bool ReadBufferExporter(unsigned int inode, std::string* exporter,
                         const std::string& dmabuf_sysfs_path = "/sys/kernel/dmabuf/buffers");
 
 /* Reads the size of the DMA buffer with @inode */
-bool ReadBufferSize(unsigned int inode, unsigned int* size,
+bool ReadBufferSize(unsigned int inode, uint64_t* size,
                     const std::string& dmabuf_sysfs_path = "/sys/kernel/dmabuf/buffers");
 
 }  // namespace dmabufinfo
